Added new readme images and refactored cn to root lib/utils
